+++ demo/sovity-messenger-demo/src/main/java/de/sovity/edc/extension/sovitymessenger/demo/message/Sqrt.java
@@ -11,7 +11,8 @@
  *       sovity GmbH - initial API and implementation
  */
 
-package de.sovity.edc.extension.messenger.dto;
+package de.sovity.edc.extension.sovitymessenger.demo.message;
+
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import de.sovity.edc.extension.messenger.SovityMessage;
@@ -19,15 +20,19 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-@Getter
 @AllArgsConstructor
 @NoArgsConstructor
-public class Answer implements SovityMessage {
+@Getter
+public class Sqrt implements SovityMessage {
+
+    private static final String TYPE = "sqrt";
+
     @Override
     public String getType() {
-        return getClass().getCanonicalName();
+        return TYPE;
     }
 
     @JsonProperty
-    private int answer;
+    private double value;
+
 }
