+++ demo/sovity-messenger-demo/src/main/java/de/sovity/edc/extension/sovitymessenger/demo/message/UnregisteredMessage.java
@@ -11,23 +11,14 @@
  *       sovity GmbH - initial API and implementation
  */
 
-package de.sovity.edc.extension.messenger.dto;
+package de.sovity.edc.extension.sovitymessenger.demo.message;
 
-import com.fasterxml.jackson.annotation.JsonProperty;
 import de.sovity.edc.extension.messenger.SovityMessage;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
 
-@Getter
-@AllArgsConstructor
-@NoArgsConstructor
-public class Answer implements SovityMessage {
+public class UnregisteredMessage implements SovityMessage {
+
     @Override
     public String getType() {
-        return getClass().getCanonicalName();
+        return "unregistered";
     }
-
-    @JsonProperty
-    private int answer;
 }
