+++ demo/sovity-messenger-demo/src/test/resources/META-INF/services/org.eclipse.edc.spi.system.ServiceExtension
