--- extensions/sovity-messenger-demo/src/main/java/de/sovity/edc/extension/sovitymessenger/demo/message/Common.java
@@ -1,18 +1,0 @@
-/*
- *  Copyright (c) 2024 sovity GmbH
- *
- *  This program and the accompanying materials are made available under the
- *  terms of the Apache License, Version 2.0 which is available at
- *  https://www.apache.org/licenses/LICENSE-2.0
- *
- *  SPDX-License-Identifier: Apache-2.0
- *
- *  Contributors:
- *       sovity GmbH - initial API and implementation
- */
-
-package de.sovity.edc.extension.sovitymessenger.demo.message;
-
-public interface Common {
-    String ROOT = "de.sovity.demo.calculator.";
-}
