--- extensions/sovity-messenger/src/main/java/de/sovity/edc/extension/messenger/api/SovityMessageApi.java
@@ -1,18 +1,0 @@
-/*
- *  Copyright (c) 2024 sovity GmbH
- *
- *  This program and the accompanying materials are made available under the
- *  terms of the Apache License, Version 2.0 which is available at
- *  https://www.apache.org/licenses/LICENSE-2.0
- *
- *  SPDX-License-Identifier: Apache-2.0
- *
- *  Contributors:
- *       sovity GmbH - initial API and implementation
- */
-
-package de.sovity.edc.extension.messenger.api;
-
-public interface SovityMessageApi {
-    String PATH = "/sovity/message/generic";
-}
