--- extensions/sovity-messenger/src/main/java/de/sovity/edc/extension/messenger/api/SovityMessengerException.java
@@ -1,32 +1,0 @@
-/*
- *  Copyright (c) 2024 sovity GmbH
- *
- *  This program and the accompanying materials are made available under the
- *  terms of the Apache License, Version 2.0 which is available at
- *  https://www.apache.org/licenses/LICENSE-2.0
- *
- *  SPDX-License-Identifier: Apache-2.0
- *
- *  Contributors:
- *       sovity GmbH - initial API and implementation
- */
-
-package de.sovity.edc.extension.messenger.api;
-
-public class SovityMessengerException extends RuntimeException {
-    public SovityMessengerException(String message) {
-        super(message);
-    }
-
-    public SovityMessengerException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-    public SovityMessengerException(Throwable cause) {
-        super(cause);
-    }
-
-    public SovityMessengerException(String message, Throwable cause, boolean enableSuppression, boolean writableStackTrace) {
-        super(message, cause, enableSuppression, writableStackTrace);
-    }
-}
