+++ extensions/sovity-messenger/src/main/java/de/sovity/edc/extension/messenger/impl/Handler.java
@@ -13,15 +13,7 @@
 
 package de.sovity.edc.extension.messenger.impl;
 
-import lombok.AllArgsConstructor;
-import lombok.Getter;
+import java.util.function.Function;
 
-@AllArgsConstructor
-@Getter
-public enum SovityMessengerStatus {
-
-    NO_HANDLER("no_handler"),
-    OK("ok");
-
-    private String code;
+public record Handler<IN, OUT>(Class<IN> clazz, Function<IN, OUT> handler) {
 }
