--- extensions/sovity-messenger/src/main/java/de/sovity/edc/extension/messenger/package-info.java
@@ -1,5 +1,0 @@
-// TODO later
-//  https://spotbugs.readthedocs.io/en/latest/gradle.html
-//@NonNullApi
-//@NonNullFields
-//package de.sovity.edc.extension.custommessages;
