+++ extensions/sovity-messenger/src/test/java/de/sovity/edc/extension/messenger/controller/Payload.java
@@ -11,7 +11,7 @@
  *       sovity GmbH - initial API and implementation
  */
 
-package de.sovity.edc.extension.messenger.dto;
+package de.sovity.edc.extension.messenger.controller;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import de.sovity.edc.extension.messenger.SovityMessage;
@@ -22,12 +22,12 @@ import lombok.NoArgsConstructor;
 @Getter
 @AllArgsConstructor
 @NoArgsConstructor
-public class Answer implements SovityMessage {
+class Payload implements SovityMessage {
+    @JsonProperty
+    private Integer integer;
+
     @Override
     public String getType() {
-        return getClass().getCanonicalName();
+        return "payload";
     }
-
-    @JsonProperty
-    private int answer;
 }
