+++ extensions/sovity-messenger/src/test/java/de/sovity/edc/extension/messenger/dto/Addition.java
@@ -14,7 +14,7 @@
 package de.sovity.edc.extension.messenger.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-import de.sovity.edc.extension.messenger.api.SovityMessage;
+import de.sovity.edc.extension.messenger.SovityMessage;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
