+++ extensions/sovity-messenger/src/test/java/de/sovity/edc/extension/messenger/dto/Multiplication.java
@@ -14,13 +14,11 @@
 package de.sovity.edc.extension.messenger.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-import de.sovity.edc.extension.messenger.api.SovityMessage;
+import de.sovity.edc.extension.messenger.SovityMessage;
 import lombok.AllArgsConstructor;
-import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-@Builder
 @AllArgsConstructor
 @NoArgsConstructor
 @Getter
