+++ settings.gradle.kts
@@ -12,7 +12,6 @@ include(":extensions:policy-referring-connector")
 include(":extensions:policy-time-interval")
 include(":extensions:postgres-flyway")
 include(":extensions:sovity-messenger")
-include(":extensions:sovity-messenger-demo")
 include(":extensions:sovity-edc-extensions-package")
 include(":extensions:test-backend-controller")
 include(":extensions:transfer-process-status-checker")
@@ -35,6 +34,7 @@ include(":launchers:connectors:sovity-ce")
 include(":launchers:connectors:sovity-dev")
 include(":launchers:connectors:test-backend")
 include(":tests")
+include(":demo:sovity-messenger-demo")
 include(":utils:catalog-parser")
 include(":utils:json-and-jsonld-utils")
 include(":utils:test-connector-remote")
